# lib/api-zod/src/generated/types/generationJob.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GenerationJobProvider } from './generationJobProvider';
import type { GenerationJobStatus } from './generationJobStatus';

export interface GenerationJob {
  id: number;
  productModelId: number;
  status: GenerationJobStatus;
  provider: GenerationJobProvider;
  attempt: number;
  maxAttempts: number;
  /** @nullable */
  lastError?: string | null;
  createdAt: Date;
  updatedAt: Date;
}