# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProductDimensions } from './productDimensions';
import type { ProductSource } from './productSource';

export interface Product {
  id: number;
  /** @nullable */
  shopifyProductId?: string | null;
  title: string;
  handle: string;
  category: string;
  eligible: boolean;
  eligibilityReason: string;
  /** @nullable */
  primaryImageUrl?: string | null;
  imageUrls: string[];
  variantCount: number;
  dimensions?: ProductDimensions | null;
  /** 'sample' rows are seeded demo data; 'shopify' rows come from a live store sync. */
  source: ProductSource;
  createdAt: Date;
  updatedAt: Date;
}