# lib/api-zod/src/generated/types/productModel.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProductModelProvider } from './productModelProvider';
import type { ProductModelStatus } from './productModelStatus';

export interface ProductModel {
  id: number;
  productId: number;
  status: ProductModelStatus;
  provider: ProductModelProvider;
  /** @nullable */
  providerTaskId?: string | null;
  /** @nullable */
  originalModelUrl?: string | null;
  /** @nullable */
  optimizedModelUrl?: string | null;
  /** @nullable */
  thumbnailUrl?: string | null;
  /** @nullable */
  imageSetHash?: string | null;
  /** @nullable */
  fileSizeBytes?: number | null;
  sourceImageCount: number;
  /** @nullable */
  errorMessage?: string | null;
  /** @nullable */
  reviewer?: string | null;
  /** @nullable */
  reviewedAt?: Date | null;
  /** @nullable */
  rejectionReason?: string | null;
  /** @nullable */
  approvedAt?: Date | null;
  /** @nullable */
  publishedAt?: Date | null;
  createdAt: Date;
  updatedAt: Date;
}