# lib/api-zod/src/generated/types/providerStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProviderStatusMode } from './providerStatusMode';

export interface ProviderStatus {
  configured: boolean;
  mode: ProviderStatusMode;
  detail: string;
}